+++ NEWS.md
@@ -3,6 +3,7 @@
 ## New features
 
 * Added extra helper functions to extract all problem instance-specific data from MPM2 generator
+* Added R-based evaluation environment for MPM2 generator which greatly accelerates evaluation in multi-objective settings
 
 ## Bugfixes